[update] 길면 기차, 기차는 빨라, 빠른 것은 비행기(9493) - 런타임 에러 (ValueError) 해결

+++ baekjoon/bronze/III/길면 기차, 기차는 빨라, 빠른 것은 비행기(9493).py
@@ -1,5 +1,5 @@
 i, I = 60, 3600
 for l in [*open(0)][:-1]:
     m, a, b = map(int, l.split())
-    x = m*I/a-m*I/b
+    x = int(m*I/a-m*I/b)
     print(f"{x//I}:{x%I//i:02d}:{x%i:02d}")
